Refactor: extract tasks from refactoring pipelines
