+++ lua/refactoring/tasks/extract_method.lua
@@ -1,0 +1,96 @@
+local ts_utils = require("nvim-treesitter.ts_utils")
+local Query = require("refactoring.query")
+local utils = require("refactoring.utils")
+local get_input = require("refactoring.get_input")
+
+local M = {}
+
+local function get_local_definitions(bufnr, local_defs, function_args)
+    local local_def_map = {}
+
+    for _, def in pairs(local_defs) do
+        local_def_map[ts_utils.get_node_text(def, bufnr)[1]] = true
+    end
+
+    for _, def in pairs(function_args) do
+        local_def_map[ts_utils.get_node_text(def, bufnr)[1]] = true
+    end
+
+    return local_def_map
+end
+
+local function get_selected_local_references(refactor)
+    local function_args = refactor.query:pluck_by_capture(
+        refactor.scope,
+        Query.query_type.FunctionArgument
+    )
+    local local_def_map = get_local_definitions(
+        refactor.bufnr,
+        refactor.selected_local_defs,
+        function_args
+    )
+
+    local local_references = refactor.locals:pluck_by_capture(
+        refactor.scope,
+        Query.query_type.Reference
+    )
+    local selected_local_references = {}
+
+    for _, local_ref in pairs(local_references) do
+        local local_name = ts_utils.get_node_text(local_ref, refactor.bufnr)[1]
+        if
+            utils.range_contains_node(local_ref, refactor.region:to_ts())
+            and local_def_map[local_name]
+        then
+            selected_local_references[local_name] = true
+        end
+    end
+    return selected_local_references
+end
+
+local function extract_to_region(refactor, target_region)
+    local selected_local_references = get_selected_local_references(refactor)
+    local function_name = get_input("106: Extract Function Name > ")
+
+    local function_body = refactor.region:get_text()
+    table.insert(function_body, refactor.code["return"]("fill_me"))
+    local args = vim.fn.sort(vim.tbl_keys(selected_local_references))
+
+    local function_code = refactor.code["function"]({
+        name = function_name,
+        args = args,
+        body = function_body,
+    })
+
+    refactor.text_edits = {
+        {
+            region = target_region,
+            text = function_code,
+            bufnr = refactor.buffers[2],
+        },
+        {
+            region = refactor.region,
+            text = refactor.code.constant({
+                name = "fill_me",
+                value = refactor.code.call_function({
+                    name = function_name,
+                    args = args,
+                }),
+            }),
+        },
+    }
+
+    return true, refactor
+end
+
+function M.extract_method_to_file(refactor)
+    local target_region = utils.get_top_of_file_region(refactor.scope)
+    return extract_to_region(refactor, target_region)
+end
+
+function M.extract_method(refactor)
+    local target_region = utils.region_above_node(refactor.scope)
+    return extract_to_region(refactor, target_region)
+end
+
+return M
