+++ lua/refactoring/tasks/inline_variable.lua
@@ -1,0 +1,4 @@
+local function inline_variable(refactor)
+    return true, refactor
+end
+return inline_variable
